+++ frontend/src/hooks/index.ts
@@ -6,3 +6,5 @@ export { default as useTimeout } from './useTimeout';
 export { default as useRouter } from './useRouter';
 export { default as useForm } from './useForm';
 export { default as useWorkbook } from './useWorkbook';
+export { default as usePublicWorkbook } from './usePublicWorkbook';
+export { default as usePublicCard } from './usePublicCard';
